[Feat] #2 : logout, jwt_required, user_lookup_loader
When logging out a user, simply use the unset_jwt_cookies function.
By using jwt_required(), a route can be protected so that only
authenticated users have access to it.
When using jwt_required, it is necessary to define user_lookup_loader.
It finds the user from the database based on the identity information
saved in the jwt token.