[Feat] #2 : save rsfh_tkn in db and function to update access_tkn
When the user signs in, save the created refresh token to the database.
Also, set the refresh token expiration time in the config file.
Create a expired_token_loader function that will be called when the
access token received from the client is expired.
Using the information inside the expired access token,
retrieve the user information from the database.
Check whether the refresh token is valid
by  using the decode_token function.
It returns the decoded token if it is valid, otherwise an error is raised.
If no error was raised, create a new access_token and set the cookie and return.