[Feat] #5 : require and goBack in Post form
Add required attribute to the inputs in the post form.
When the user tries to submit the form with empty forms,
it will not be submitted.
Pass the history object in order to be able to use history.goBack().
If cancel button is clicked, the application will go back to the
previous page.