[Feat] #5 : use validators in signup and signin form
Manage the validation state using useState.
Check validation status using the created validators.
For this, create a new onInput function that checks
if the field values are validated and updates them in the state
every time the user changes the value of the input.
Before submitting the form, check the validation status.
If one of them is false, show the message to the user.