[Feat] #12 : style button
Instead of using bulma button, a personally styled button component was created.
This is because bulma's button was hard to properly place for unknown
css placement properties.
When using two buttons, and wanting to center them,
only the first button was centered and the second button not.
Therefore, a new button was created.