[Feat] #12 : create comment list and item
CommentList receives the list of comments from the Comment component.
Since recent comments must be shown first, reverse the array.
The reverse function reverses an array in place.
Therefore, create an empty array, place all elements of comments inside it,
and then use the reverse function.
Finally, use map to create CommentItem component for each of the comment.
The Comment component will show the respective information.
The edit and delete buttons will only be shown to the users who wrote the comments.