[Feat] #12 : delete comment
First check whether the user trying to delete the comment
is the same user who wrote the comment.
If it is, delete the comment using SP.
Instead of actually deleting the comment,
the DELETED column will be filled with the current date and time,
which is indicate the date and time when the user decided to delete it.