[Feat] #12 : edit function for CommentItem
When the user clicks the edit button of a comment,
retrieve the comment information and fill the comment form with it.
Instead of using the data in the browser, the body content is retrieved from the server
because the user might change the comment from one browser and
change it again in another browser.
Therefore, use the database's data to update the comment.