[Refactor] : remove useCallback
useCallback was used to solve the too many re-renders problem.
However, the problem was that the dispatch was called when the component was rendered,
but what was intended was to be called at a specific point of time (after some user action).
The problem was solved by wrapping the dispatch function with another function.
This way, it will not run at render time, but when the declared function is called.