[Review] : db settings for dictionary cursor
It is possible to set the db to return the results in dictionary format.
However, by using the settings, the column names are returned with the
case as it is. Currently, the db column names are capitalized,
but the client and server uses them in the lower case.
Because too many changes are required, do not add the dict cursor settings.