[Review] : create a single function for api call
Instead of creating a function for each type of api call,
create a single one that runs the axios function with different conditions
given the apiType, id, and/or the data.
Also, create a separate paths file which will contain all the paths
needed in the project.
Create two different paths object: one for api calls, one for history.