[Fix]: password2 validation not working, remove useValidation
password2 validation in useValidation hook was not correctly working.
This was because in order to do so, the useValidation hook needed to use
the state of the actual value of the inputs.
However, since the function is run by user input event, it is hard to pass
the state to the function. Therefore, instead of using hook,
the validation function was declared inside the component, where
the state of the inputs is reachable.